Move CountedPtbBuilder logic inside SitePtb.
As we already need to think about the count there to check multiple move-calls happen atomically.
